shared_infra/pyproject.toml
[tool.poetry]
name = "shared-infra"
version = "0.1.0"
description = ""
authors = ["Simon Coope <sjcoope@amazon.com>"]
readme = "README.md"
packages = [{include = "shared_infra"}]

[tool.poetry.dependencies]
python = "^3.9"
aws-cdk-lib = "^2.49.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
